Use JWT in addition to the GitHub OAuth2 application-based authentication
This allows the tidycat backend to validate requests using a mechanism
(JWT) that is independent of the primary authentication mechanism
(GitHub).

Read more about it at: https://disjoint.ca/til/2016/03/22/chaining-authentication-in-an-ember-app

Right now, the mock server backend responds on two endpoints:

- `/api/auth/token`
- `/api/auth/refresh`

The `token` endpoint hands out a new JWT token, whereas the `refresh`
endpoint validates the previous JWT and then hands out a new one.

The idea here is to keep this mock backend in place until I get the real
one up and running!