Use a cleaner method for chaining ESA authentication
Since ESA triggers the sessionAuthenticated method post-authentication,
that place seemed like a reasonable choice to trigger the JWT
authentication.

Without this, there was a race condition between authentication
finishing and the session object being ready to use, which resulted in
the model (in the "authenticatedRoute") being in a bad state.