Inject the model directly from the application route
This allows the model to be used by all child routes without too much
extra effort.

Note that the `_this.refresh();` call is because the model needs the
`session` object before it is valid, and this refresh takes care of
refreshing the model before the transition.

This is quite hacky, meh.