Add an Ember component to display logged-in-context specific buttons in the navbar
Using a component vs adding this directly to the application.hbs file
makes it a lot easier to isolate and test.

There *appears* to be a bug in where the Font Awesome icons for 'Alerts'
and 'Sign out' doesn't appear to render until the dom is refreshed (page
size changes, etc). The odd thing here is that this only manifests
itself in Chrome and not Firefox.

I'm going to assume that is an anomaly of my environment and let that
be, for now.