Improve the unit and acceptance testing
One of big problems I ran into while attempting to mock out ember data
and other http calls was Ember.run.later tasks.

A big part of the ember-simple-auth-token library is the ability to
check for refresh tokens. The reason for turning this off in testing is
because attempting to mock out run.later is quite harrowing and it's
probably best to disable this in testing entirely.

```
ENV['ember-simple-auth-token'].refreshAccessTokens = false;
```